requirements.txt
asgiref==3.6.0
boto3==1.26.63
botocore==1.29.63
crispy-bootstrap5==0.7
dj-database-url==0.5.0
Django==4.1.7
django-allauth==0.52.0
django-countries==7.5.1
django-crispy-forms==2.0
django-storages==1.13.2
djlint==1.19.12
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
sqlparse==0.4.3